Refactor configuration management and remove constants module
- Introduced environment variable loading and path management directly in config module.
- Removed the constants.py file and integrated its functionality into config initialization.
- Added initialization logic for configuration directories and files.
- Updated database URI handling and paths for local storage.
- Adjusted imports in db module to reflect new config structure.
- Cleaned up unnecessary whitespace in various files.
