+++ wembed/__init__.py
@@ -1,0 +1,2 @@
+import config
+import cli
