+++ wembed/db/__init__.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import psycopg2
 import typer
-from wembed.config import app_config
+from config import app_config
 from sqlalchemy import Engine, create_engine, text
 from sqlalchemy.orm import Session, sessionmaker
